news-analyzer/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
playwright==1.40.0
aiohttp==3.9.1
feedparser==6.0.10
redis==5.0.1
textblob==0.17.1
transformers==4.36.2
torch==2.1.2
apscheduler==3.10.4
asyncio
python-dateutil==2.8.2
lxml==4.9.3
html5lib==1.1
fake-useragent==1.4.0
requests-html==0.10.0
vaderSentiment==3.3.2
fastapi
uvicorn
